Generate report and improve argument parsing
Now the argument parsing is done using the argparse library. The program
accepts an input file, an optional output file (for the output spell)
and an optional report file. When the output file is not specified, the
spell is printed in the standard output. When the report file is not
specified (and the spell has not been printed to the standard output),
the report will be printed to the standard output.