Fix file download
The download URL (e.g. /uploads/file.zip) was only based on the path of
the file (/uploads/file.zip). Because of this, if the upload folder
changes, it will return a 404 when we try to download the file. This
commit fixes that problem by returning always the same url:
uploads/FILE, where FILE is the only variable and does not include the
path to the upload folder.

Also, the default upload folder changes to the /tmp directory.