Render steps
Before, the template expected a list of controls to render. Now, the
input is a list of steps, which contains the controls. The controls now
have a type property again so that we can render it correctly.

Also, the template is modularized, extracting the code that renders the
controls into it's own file (spel/control.spl).